Update CSVDataContainer.java
putDataSetRows: Use FileUtil.getRowsAsList instead of creating an own Reader instance.

Reason: The readLine method of the BufferedReader has to be surrounded by an IOException. But the private method only throwed IOException where it got passed into a RuntimeException in the calling public method. The FileUtil methods are public and handle the IOExceptions of the Reader directly and no error occurs.