Change behaviour when inserting fields into CSVDataContainer
The values have been cleaned in the past, before adding a row into a CSVDataContainer. Deleting the starting and ending quotes, and set the rows array entry to null for empty fields. This function is now deactivated, so that the field content will be retrieved as it is in the future.